Rename collections -> collection_helpers to avoid conflict with built-in module.
